requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.9.0
cssselect2==0.7.0
datasets==3.6.0
decorator==5.1.1
dill==0.3.8
Django==5.1.5
django-cors-headers==4.6.0
django-celery-results==2.5.1
django-filter==24.3
django-weasyprint==2.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
evaluate==0.4.3
exceptiongroup==1.2.2
filelock==3.17.0
fonttools==4.55.6
frozenlist==1.6.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
jiwer==3.1.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain-core==0.3.31
langchain-ollama==0.2.2
langgraph==0.2.67
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.1
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
Markdown==3.7
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
mysqlclient==2.2.7
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.7
faster-whisper==1.1.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.50
propcache==0.3.1
psycopg2-binary==2.9.10
pyarrow==20.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
xxhash==3.5.0
yarl==1.20.0
zopfli==0.2.3.post1
zstandard==0.23.0
pyannote.audio==3.3.2
git+https://github.com/huggingface/diarizers.git
google-generativeai==0.8.6
